gmail-mcp/requirements.txt
fastmcp
google-auth
google-auth-oauthlib
google-api-python-client
pinecone-client
langchain-pinecone
langchain-nomic
langchain-groq
python-dotenv
pytest 